requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytest==8.3.3
pytest-cov==6.0.0
